+++ dashboard/src/app/modules/administration/pages/certificate-list/certificate-list.component.html
@@ -1,5 +1,6 @@
 <div>
   <div class="FilterTop">
+    <h3>Attestation</h3>
     <app-operator-autocomplete [parentForm]="filterForm"></app-operator-autocomplete>
   </div>
 
@@ -33,7 +34,7 @@
 
     <ng-container matColumnDef="actions">
       <th mat-header-cell *matHeaderCellDef></th>
-      <td mat-cell *matCellDef="let row">
+      <td mat-cell *matCellDef="let row" class="item-actions">
         <a class="link" (click)="download(row, 'application/pdf')">
           <mat-icon color="primary" matTooltip="Télécharger la version PDF">picture_as_pdf</mat-icon>
         </a>
