+++ dashboard/src/app/modules/certificate/certificate.module.ts
@@ -11,7 +11,7 @@ import { GenerateComponent } from './pages/generate/generate.component';
 import { CheckComponent } from './pages/check/check.component';
 import { CertificateLayoutComponent } from './certificate-layout/certificate-layout.component';
 import { CertificateRoutingModule } from './certificate-routing.module';
-import { CertificateListComponent } from './pages/certificate-list/certificate-list.component';
+import { CertificateListComponent } from '../administration/pages/certificate-list/certificate-list.component';
 import { TerritoryUiModule } from '../territory/modules/territory-ui/territory-ui.module';
 import { OperatorUiModule } from '../operator/modules/operator-ui/operator-ui.module';
